cleanup(tests): extract engine fixture
